fix(node): resolve heartbeat compilation — type cast + match arm
- node.rs: cast total_atoms u32→usize for fragments_processed
- handler.rs: add Heartbeat match arm (handled in Node::process())
- 38 Rust + 15 Python smoke tests GREEN

+++ bizra-omega/bizra-node/src/node.rs
@@ -430,7 +430,7 @@ impl Node {
         // Check if there are pending fragments from recent RECEIVE commands.
         // In MVP, this counts accumulated atoms since last heartbeat.
         let summary = self.runtime.pipeline().knowledge_summary();
-        report.fragments_processed = summary.total_atoms;
+        report.fragments_processed = summary.total_atoms as usize;
 
         // ── Loop B: Memory → Cognition ──────────────────────
         // Trigger synthesis if enough time has elapsed.
